\name{linBilin-package}
\alias{linBilin-package}
\alias{linBilin}
\docType{package}
\title{
Fast linear interpolation of 1-, 2- and 3-dimensional functions
}

\details{
\tabular{ll}{
Package: \tab linBilin\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-09-08\cr
License: \tab What license is it under?\cr
}

Given a function defined on one, two, or three variables, this package will provide very fast linear interpolation using the Armadillo C++ library.  A C++ interface is also available in the \code{linBillin.hpp} file in the \code{src/} directory.

}
\author{
Philip Barrett

Maintainer: Philip Barrett <pobarrett@gmail.com>

}

\keyword{ package, interpolation, approximation, linear }
\seealso{
\code{\link{lin}}
\code{\link{bilin}}
\code{\link{trilin}}
}
\examples{
# See individual function documentation for examples
}
